Fix problems that were preventing some (deprecated) SCons build targets from compiling/linking
These changes allow the cgi, lowtools, and shell build targets to fully compile/link. Figured I might as well get them working for posterity, if I wasn't going to remove them as options in the SCons build.

+++ code/code/cgi/session.cgi.cc
@@ -235,7 +235,7 @@ sstring TSession::generateSessionID() {
   seed[0] = time(NULL);
   seed[1] = random();
   seed[2] = getpid();
-  seed[3] = (int)&seed;
+  seed[3] = static_cast<int>((long)&seed);
 
   int c = 0;
   for (int i = 0; i < 4; ++i) {
